+++ packages/react/src/slider/slider-thumb.tsx
@@ -6,24 +6,16 @@ import { useSliderContext } from './slider-context'
 export type SliderThumbProps = ComponentPropsWithoutRef<typeof ark.div>
 
 export const SliderThumb = forwardRef<HTMLDivElement, SliderThumbProps>((props, ref) => {
-  const { thumbProps } = useSliderContext()
+  const { thumbProps, hiddenInputProps } = useSliderContext()
   const mergedProps = mergeProps(thumbProps, props)
   return (
-    <ark.div {...mergedProps} ref={ref}>
-      <SliderHiddenInput />
-      {props.children}
-    </ark.div>
+    <>
+      <ark.div {...mergedProps} ref={ref}>
+        {props.children}
+      </ark.div>
+      <input {...hiddenInputProps} />
+    </>
   )
 })
 
 SliderThumb.displayName = 'SliderThumb'
-
-const SliderHiddenInput = forwardRef<HTMLInputElement, ComponentPropsWithoutRef<typeof ark.input>>(
-  (props, ref) => {
-    const { hiddenInputProps } = useSliderContext()
-    const mergedProps = mergeProps(hiddenInputProps, props)
-    return <ark.input {...mergedProps} ref={ref} />
-  },
-)
-
-SliderHiddenInput.displayName = 'SliderHiddenInput'