+++ packages/solid/src/checkbox/checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import { createSignal } from 'solid-js'
 import type { Meta } from 'storybook-solidjs'
-import { Checkbox, CheckboxControl, CheckboxInput, CheckboxLabel, type CheckedState } from './'
+import { Checkbox, CheckboxControl, CheckboxLabel, type CheckedState } from './'
 import './checkbox.css'
 
 const meta: Meta = {
@@ -12,7 +12,6 @@ export default meta
 export const Basic = () => (
   <Checkbox>
     <CheckboxLabel>Checkbox</CheckboxLabel>
-    <CheckboxInput />
     <CheckboxControl />
   </Checkbox>
 )
@@ -23,7 +22,6 @@ export const Controlled = () => {
     <>
       <Checkbox checked={checked()} onChange={(e) => setChecked(e.checked)}>
         <CheckboxLabel>Checkbox</CheckboxLabel>
-        <CheckboxInput />
         <CheckboxControl />
       </Checkbox>
     </>
@@ -33,7 +31,6 @@ export const Controlled = () => {
 export const Indeterminate = () => (
   <Checkbox checked="indeterminate">
     <CheckboxLabel>Checkbox</CheckboxLabel>
-    <CheckboxInput />
     <CheckboxControl />
   </Checkbox>
 )
@@ -43,7 +40,6 @@ export const RenderProp = () => (
     {(api) => (
       <>
         <CheckboxLabel>Checkbox</CheckboxLabel>
-        <CheckboxInput />
         <CheckboxControl>
           {api().isChecked && <span>✓</span>}
           {api().isIndeterminate && <span>-</span>}