+++ packages/solid/src/tags-input/tags-input-control.tsx
@@ -6,6 +6,12 @@ export type TagsInputControlProps = HTMLArkProps<'div'>
 
 export const TagsInputControl = (props: TagsInputControlProps) => {
   const api = useTagsInputContext()
-  const controlProps = mergeProps(() => api().controlProps, props)
-  return <ark.div {...controlProps} />
+  const mergedProps = mergeProps(() => api().controlProps, props)
+
+  return (
+    <>
+      <ark.div {...mergedProps} />
+      <input {...api().hiddenInputProps} />
+    </>
+  )
 }