+++ packages/vue/src/checkbox/stories/controlled.stories.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import { Checkbox, CheckboxControl, CheckboxInput, CheckboxLabel } from '..'
+import { Checkbox, CheckboxControl, CheckboxLabel } from '..'
 import '../checkbox.css'
 
 const checked = ref(true)
@@ -8,7 +8,6 @@ const checked = ref(true)
 <template>
   <Checkbox v-model="checked">
     <CheckboxLabel>Checkbox</CheckboxLabel>
-    <CheckboxInput />
     <CheckboxControl />
   </Checkbox>
 </template>