+++ packages/vue/src/checkbox/stories/indeterminate.stories.vue
@@ -1,10 +1,9 @@
 <script setup lang="ts">
 import { reactive, ref, watch } from 'vue'
-import { Checkbox, CheckboxControl, CheckboxInput, CheckboxLabel } from '..'
+import { Checkbox, CheckboxControl, CheckboxLabel, type CheckedState } from '..'
+import '../checkbox.css'
 import CheckIcon from './CheckIcon.vue'
 import MinusIcon from './MinusIcon.vue'
-import '../checkbox.css'
-import type { CheckedState } from '@zag-js/checkbox/dist/checkbox.types'
 
 const parentChecked = ref<CheckedState>(false)
 const childCheckedItems = reactive([false, false])
@@ -34,7 +33,6 @@ watch(parentChecked, (parentVal) => {
         <MinusIcon v-else-if="parentChecked === 'indeterminate'" />
       </CheckboxControl>
       <CheckboxLabel> Parent Checkbox</CheckboxLabel>
-      <CheckboxInput />
     </Checkbox>
     <div style="padding-left: 20px; display: flex; flex-direction: column; gap: 4px">
       <Checkbox v-model="childCheckedItems[0]">
@@ -42,14 +40,12 @@ watch(parentChecked, (parentVal) => {
           <CheckIcon v-if="childCheckedItems[0]" />
         </CheckboxControl>
         <CheckboxLabel> Child One Checkbox</CheckboxLabel>
-        <CheckboxInput data-testid="child-one-input" />
       </Checkbox>
       <Checkbox v-model="childCheckedItems[1]">
         <CheckboxControl>
           <CheckIcon v-if="childCheckedItems[1]" />
         </CheckboxControl>
         <CheckboxLabel>Child Two Checkbox</CheckboxLabel>
-        <CheckboxInput data-testid="child-two-input" />
       </Checkbox>
     </div>
   </div>