+++ packages/vue/src/tags-input/tags-input-control.tsx
@@ -10,10 +10,12 @@ export const TagsInputControl = defineComponent({
     const api = useTagsInputContext()
 
     return () => (
-      <ark.div {...api.value.controlProps} {...attrs}>
-        {slots.default?.()}
+      <>
+        <ark.div {...api.value.controlProps} {...attrs}>
+          {slots.default?.()}
+        </ark.div>
         <input {...api.value.hiddenInputProps} />
-      </ark.div>
+      </>
     )
   },
 })